mrc-kv-cache-configurator-core(kv-pool-runtime-move,mechanical_provable): Move is_post_capture_kv_active / PostCaptureKVResize / compute_post_capture_kv_resize to model_runner_components/kv_pool_runtime.py (cut+paste)
